refactor(backend): move router-inline models into schemas/
bpm.py, soundcloud.py and system_playlists.py declared their request/response
models inline while every other router imports from schemas/. They now follow
the same convention via schemas/bpm.py and schemas/soundcloud.py.

The planned schemas/api vs schemas/config split is dropped: ai.py, ruleset.py
and profile_group.py each mix persisted-settings models with wire DTOs inside
one module, so the boundary does not fall between modules. Grouping by feature
is the better fit — see BACKEND-RESTRUCTURE.md.
